Delete utils-csv. Change init database to service.
