+++ openpkw-rest/src/test/java/org/openpkw/web/config/TestAppConfig.java
@@ -8,7 +8,7 @@ import org.springframework.context.annotation.Configuration;
  * @author Sebastian Pogorzelski
  */
 @Configuration
-@ComponentScan(basePackages = {"org.openpkw.qr", "org.openpkw.services", "org.openpkw.web"})
+@ComponentScan(basePackages = {"org.openpkw.services.qr", "org.openpkw.services", "org.openpkw.web"})
 public class TestAppConfig {
 
 }
