--- openpkw-services/src/main/java/org/openpkw/qr/dto/ProtocolDTO.java
@@ -1,10 +1,0 @@
-package org.openpkw.qr.dto;
-
-/**
- * Protocol DTO
- * @author Sebastian Pogorzelski
- */
-public class ProtocolDTO {
-
-
-}
