--- openpkw-services/src/main/java/org/openpkw/qr/parser/QrParserService.java
@@ -1,16 +1,0 @@
-package org.openpkw.qr.parser;
-
-import org.springframework.stereotype.Service;
-
-/**
- * Service for parsing qr
- * @author Sebastian Pogorzelski
- */
-@Service
-public class QrParserService {
-
-    public QrWrapper parse(String qr) {
-        return new DefaultQrWrapper(qr);
-    }
-
-}
