--- openpkw-services/src/main/java/org/openpkw/qr/services/QrService.java
@@ -1,14 +1,0 @@
-package org.openpkw.qr.services;
-
-import org.openpkw.qr.dto.QrDTO;
-import org.openpkw.qr.dto.QrResultDTO;
-
-/**
- * Service for parsing and saving QR data from mobile
- * @author Sebastian Pogorzelski
- */
-public interface QrService {
-
-    QrResultDTO saveResult(QrDTO qrDTO);
-
-}
