+++ openpkw-services/src/main/java/org/openpkw/services/init/InitService.java
@@ -1,0 +1,8 @@
+package org.openpkw.services.init;
+
+/**
+ * Created by mrozi on 12.01.16.
+ */
+public interface InitService {
+    void initDatabase();
+}
