+++ openpkw-services/src/main/java/org/openpkw/services/qr/QrService.java
@@ -1,0 +1,14 @@
+package org.openpkw.services.qr;
+
+import org.openpkw.services.qr.dto.QrDTO;
+import org.openpkw.services.qr.dto.QrResultDTO;
+
+/**
+ * Service for parsing and saving QR data from mobile
+ * @author Sebastian Pogorzelski
+ */
+public interface QrService {
+
+    QrResultDTO saveResult(QrDTO qrDTO);
+
+}
