+++ openpkw-services/src/main/java/org/openpkw/services/qr/dto/ProtocolDTO.java
@@ -1,0 +1,10 @@
+package org.openpkw.services.qr.dto;
+
+/**
+ * Protocol DTO
+ * @author Sebastian Pogorzelski
+ */
+public class ProtocolDTO {
+
+
+}
