+++ openpkw-services/src/main/java/org/openpkw/services/qr/parser/QrIndex.java
@@ -1,0 +1,48 @@
+package org.openpkw.services.qr.parser;
+
+/**
+ * Enum contains indexes in QR
+ * @author Sebastian Pogorzelski
+ */
+enum QrIndex {
+
+    TERRITORIAL_CODE(0),
+    PERIPHERY_NUMBER(1),
+    DISTRICT_NUMBER(2),
+
+    VOTINGCARDS_TOTALENTITLEDTOVOTE(3),
+    VOTINGCARDS_TOTALCARDS(4),
+    VOTINGCARDS_UNUSEDCARDS(5),
+    VOTINGCARDS_REGULARVOTERS(6),
+    VOTINGCARDS_REPRESENTATIVEVOTERS(7),
+    VOTINGCARDS_CERTIFICATEVOTERS(8),
+
+    CORRESPONDENCEVOTING_ISSUEDPACKAGES(9),
+    CORRESPONDENCEVOTING_RECEIVEDREPLYENVELOPES(10),
+    CORRESPONDENCEVOTING_MISSINGSTATEMENT(11),
+    CORRESPONDENCEVOTING_MISSINGSIGNATUREONSTATEMENT(12),
+    CORRESPONDENCEVOTING_MISSINGENVELOPEFORVOTINGCARD(13),
+    CORRESPONDENCEVOTING_UNSEALEDENVELOPE(14),
+    CORRESPONDENCEVOTING_ENVELOPESTHROWNTOBALLOTBOX(15),
+
+    VOTINGCARDS_CARDSFROMBALLOTBOX(16),
+    VOTINGCARDS_CARDSFROMENVELOPES(17),
+    VOTINGCARDS_INVALIDCARDS(18),
+    VOTINGCARDS_VALIDCARDS(19),
+    VOTINGCARDS_INVALIDVOTES(20),
+    VOTINGCARDS_VALIDVOTES(21),
+
+
+    CANDIDATES_START(23);
+
+    private int index;
+
+    QrIndex(int index) {
+        this.index = index;
+    }
+
+    public int getIndex() {
+        return index;
+    }
+
+}
