+++ openpkw-services/src/main/java/org/openpkw/services/qr/parser/QrParseService.java
@@ -1,0 +1,8 @@
+package org.openpkw.services.qr.parser;
+
+/**
+ * Created by mrozi on 12.01.16.
+ */
+public interface QrParseService {
+    QrWrapper parse(String qr);
+}
