--- openpkw-services/src/main/java/org/openpkw/services/seatsinparliament/ElectedDeputyInConstituency.java
@@ -1,11 +1,0 @@
-package org.openpkw.services.seatsinparliament;
-
-public class ElectedDeputyInConstituency {
-
-    public Integer numberConstituency;
-    public String politicalParty;
-    public String deputyName;
-    public String deputySurname;
-    public Integer numberOfVotes;
-
-}
