+++ openpkw-services/src/main/java/org/openpkw/services/seatsinparliament/SeatsInParliamentCalculatorService.java
@@ -4,6 +4,6 @@ import java.util.Map;
 
 public interface SeatsInParliamentCalculatorService {
 
-    public Map<String, Integer> dHondt(int mandateInConstituency, Map<String, Integer> politicalPartyVoices);
+    Map<String, Integer> dHondt(int mandateInConstituency, Map<String, Integer> politicalPartyVoices);
 
 }
