+++ openpkw-services/src/main/resources/peripherals.csv
