--- openpkw-utils-csv/.gitignore
@@ -1,5 +1,0 @@
-/target/
-/.settings/
-.project
-.classpath
-output*.pdf
