--- openpkw-utils-csv/README.md
@@ -1,1 +1,0 @@
-Parser  pliku CSV  ( dane przekazane z PKW )
