--- openpkw-utils-csv/src/main/java/org/openpkw/utils/csv/json/Candidate.java
@@ -1,23 +1,0 @@
-package org.openpkw.utils.csv.json;
-
-public class Candidate {
-    private String candidateName;
-    private long candidateVotesNumber;
-
-    public String getCandidateName() {
-        return candidateName;
-    }
-
-    public void setCandidateName(String candidateName) {
-        this.candidateName = candidateName;
-    }
-
-    public long getCandidateVotesNumber() {
-        return candidateVotesNumber;
-    }
-
-    public void setCandidateVotesNumber(long candidateVotesNumber) {
-        this.candidateVotesNumber = candidateVotesNumber;
-    }
-
-}
