--- openpkw-utils-csv/src/main/java/org/openpkw/utils/csv/json/Comments.java
@@ -1,69 +1,0 @@
-package org.openpkw.utils.csv.json;
-
-public class Comments {
-
-    private String comments15 = "";
-    private String comments16 = "";
-    private String comments17 = "";
-    private String comments18 = "";
-    private String comments19 = "";
-    private String comments20 = "";
-    private String comments21 = "";
-
-    public String getComments15() {
-        return comments15;
-    }
-
-    public void setComments15(String comments15) {
-        this.comments15 = comments15;
-    }
-
-    public String getComments16() {
-        return comments16;
-    }
-
-    public void setComments16(String comments16) {
-        this.comments16 = comments16;
-    }
-
-    public String getComments17() {
-        return comments17;
-    }
-
-    public void setComments17(String comments17) {
-        this.comments17 = comments17;
-    }
-
-    public String getComments18() {
-        return comments18;
-    }
-
-    public void setComments18(String comments18) {
-        this.comments18 = comments18;
-    }
-
-    public String getComments19() {
-        return comments19;
-    }
-
-    public void setComments19(String comments19) {
-        this.comments19 = comments19;
-    }
-
-    public String getComments20() {
-        return comments20;
-    }
-
-    public void setComments20(String comments20) {
-        this.comments20 = comments20;
-    }
-
-    public String getComments21() {
-        return comments21;
-    }
-
-    public void setComments21(String comments21) {
-        this.comments21 = comments21;
-    }
-
-}
