--- openpkw-utils-csv/src/main/java/org/openpkw/utils/csv/json/DokumentGeneratorRequest.java
@@ -1,20 +1,0 @@
-package org.openpkw.utils.csv.json;
-
-public class DokumentGeneratorRequest {
-
-    private String templateName;
-    private PeripheryVote formData;
-
-    public DokumentGeneratorRequest(String templateName, PeripheryVote formData) {
-        this.templateName = templateName;
-        this.formData = formData;
-    }
-
-    public String getTemplateName() {
-        return templateName;
-    }
-
-    public PeripheryVote getFormData() {
-        return formData;
-    }
-}
