--- openpkw-utils-csv/src/main/java/org/openpkw/utils/csv/json/ElectoralCampaign.java
@@ -1,33 +1,0 @@
-package org.openpkw.utils.csv.json;
-
-public class ElectoralCampaign {
-
-    private String date;
-    private String startTime;
-    private String endTime;
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public String getStartTime() {
-        return startTime;
-    }
-
-    public void setStartTime(String startTime) {
-        this.startTime = startTime;
-    }
-
-    public String getEndTime() {
-        return endTime;
-    }
-
-    public void setEndTime(String endTime) {
-        this.endTime = endTime;
-    }
-
-}
