--- openpkw-utils-csv/src/main/resources/META-INF/persistence.xml
@@ -1,30 +1,0 @@
-<persistence xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
-  version="2.0" xmlns="http://java.sun.com/xml/ns/persistence">
-  <persistence-unit name="openpkw" transaction-type="RESOURCE_LOCAL">
-    <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-    <class>org.openpkw.model.entity.Candidate</class>
-    <class>org.openpkw.model.entity.DistrictCommittee</class>
-    <class>org.openpkw.model.entity.DistrictCommitteeAddress</class>
-    <class>org.openpkw.model.entity.ElectionCommittee</class>
-    <class>org.openpkw.model.entity.ElectionCommitteeDistrict</class>
-    <class>org.openpkw.model.entity.ElectionCommitteeVote</class>
-    <class>org.openpkw.model.entity.PeripheralCommittee</class>
-    <class>org.openpkw.model.entity.PeripheralCommitteeAddress</class>
-    <class>org.openpkw.model.entity.Protocol</class>
-    <class>org.openpkw.model.entity.User</class>
-    <class>org.openpkw.model.entity.UserDevice</class>
-    <class>org.openpkw.model.entity.UserSession</class>
-    <class>org.openpkw.model.entity.Vote</class>
-    
-    <properties>
-      <property name="javax.persistence.jdbc.driver" value="com.mysql.jdbc.Driver" />
-      <property name="javax.persistence.jdbc.url"  value="jdbc:mysql://localhost/openpkw?useUnicode=true&amp;characterEncoding=UTF-8" />
-      <property name="javax.persistence.jdbc.user" value="openpkw" />
-      <property name="javax.persistence.jdbc.password" value="lwejlr2k3jlsfedlk2j34" />
-      <property name="eclipselink.ddl-generation" value="drop-and-create-tables" />
-      <property name="eclipselink.ddl.default-table-suffix" value="engine=InnoDB"/>
-      <property name="eclipselink.ddl-generation.output-mode" value="database" />
-    </properties>
-  </persistence-unit>
-</persistence>
