+++ pom.xml
@@ -11,7 +11,6 @@
         <module>openpkw-services</module>
         <module>openpkw-rest</module>
         <module>openpkw-test</module>
-        <module>openpkw-utils-csv</module>
     </modules>
     <packaging>pom</packaging>
     <name>openpkw</name>
